consolidate pointing into _pointing module
