+++ romancal/orientation/_lib.py
@@ -3,21 +3,18 @@
 import dataclasses
 import logging
 import sys
-from collections import defaultdict, namedtuple
+from collections import namedtuple
 from collections.abc import Callable
-from copy import copy
 from math import cos, sin
 from typing import Any
 
 import asdf
 import numpy as np
-from astropy.table import Table
-from astropy.time import Time, TimeDelta
+from astropy.time import Time
 from stcal.alignment.util import compute_s_region_keyword
 from stcal.velocity_aberration import compute_va_effects_vector
 
-from ..lib.engdb.engdb_lib import EngDB_Value
-from ..lib.engdb.engdb_tools import engdb_service
+from . import _pointing as plib
 
 __all__ = []
 
@@ -68,15 +65,6 @@ PI2 = np.pi * 2.0
 
 # Potential errors generated by mnemonic retrival
 EXPECTED_ERRORS = (OSError, RuntimeError, ValueError)
-
-# Pointing container
-# Attributes are as follows. Except for the observation time, all values
-# are retrieved from the engineering data.
-#    fgs_q        : Quaternion representing orientation of the FGS frame relative to the Observatory frame
-#    obstime      : Time the pointing information refers to.
-#    q            : Quaternion of the FGS.
-Pointing = namedtuple("Pointing", ["fgs_q", "obstime", "q"])
-Pointing.__new__.__defaults__ = (None,) * 3
 
 
 # Transforms
@@ -150,11 +138,6 @@ class Transforms:
         asdf_file.write_to(path, all_array_storage="inline")
 
 
-# WCS reference container
-WCSRef = namedtuple("WCSRef", ["ra", "dec", "pa", "s_region"])
-WCSRef.__new__.__defaults__ = (None, None, None, "")
-
-
 @dataclasses.dataclass
 class TransformParameters:
     """Parameters required for the calculations."""
@@ -173,7 +156,7 @@ class TransformParameters:
     obsstart: float | None = None
     #: The observatory orientation, represented by the ECI quaternion,
     # and other engineering mnemonics
-    pointing: Pointing | Any = None
+    pointing: plib.Pointing | Any = None
     #: Reduction function to use on values.
     reduce_func: Callable | None = None
     #: Engineering database information
@@ -194,7 +177,7 @@ class TransformParameters:
 
     def update_from_engdb(self):
         """Update pointing information."""
-        self.pointing = get_pointing(
+        self.pointing = plib.get_pointing(
             self.obsstart,
             self.obsend,
             mnemonics_to_read=COARSE_MNEMONICS,
@@ -206,197 +189,58 @@ class TransformParameters:
     def __post_init__(self):
         # Setup the default reduction function.
         if not self.reduce_func:
-            self.reduce_func = pointing_from_average
+            self.reduce_func = plib.pointing_from_average
 
 
-def update_wcs_from_telem(model, t_pars: TransformParameters):
+# WCS reference container
+WCSRef = namedtuple("WCSRef", ["ra", "dec", "pa", "s_region"])
+WCSRef.__new__.__defaults__ = (None, None, None, "")
+
+
+def calc_attitude_matrix(wcs, yangle, position):
     """
-    Update WCS pointing information.
-
-    Given a `roman.datamodels.DataModel`, determine the simple WCS parameters
-    from the SIAF keywords in the model and the engineering parameters
-    that contain information about the telescope pointing.
-
-    It presumes all the accessed keywords are present.
+    Calculate the DCM attitude from known positions and roll angles.
 
     Parameters
     ----------
-    model : `~roman.datamodels.DataModel`
-        The model to update. The update is done in-place.
+    wcs : WCSRef
+        The sky position.
 
-    t_pars : `TransformParameters`
-        The transformation parameters. Parameters are updated during processing.
+    yangle : float
+        The IdlYangle of the point in question.
 
-    Returns
-    -------
-    transforms : Transforms or None
-        If available, the transformation matrices.
-    """
-    logger.info("Updating wcs from telemetry.")
-
-    # Initialization. If provided, provide a default Pointing.
-    transforms = None  # Assume no transforms are calculated.
-    quality = None  # Unknown pointing quality.
-
-    # Get the pointing information
-    try:
-        t_pars.update_from_engdb()
-    except ValueError as exception:
-        logger.error("Cannot retrieve valid engineering orientation data")
-        if t_pars.default_quaternion is None or not t_pars.allow_default:
-            logger.error("Use of default orientation has been disabled. Aborting.")
-            raise
-        else:
-            logger.warning("Exception is %s", exception)
-            obstime = Time(
-                (t_pars.obsstart.mjd + t_pars.obsend.mjd) / 2.0, format="mjd"
-            )
-            logger.warning(
-                "Using provided default quaternion: %s", t_pars.default_quaternion
-            )
-            logger.warning("    at time %s", obstime.iso)
-            logger.info("Setting pointing quality to PLANNED")
-            t_pars.pointing = Pointing(q=t_pars.default_quaternion, obstime=obstime)
-            quality = "PLANNED"
-    else:
-        logger.info("Successful read of engineering quaternions:")
-        logger.info("\tPointing: %s", t_pars.pointing)
-        quality = "CALCULATED"
-
-    # Attempt to calculate WCS information
-    try:
-        wcsinfo, vinfo, transforms = calc_wcs(t_pars)
-        logger.info("Setting pointing quality to %s", quality)
-    except EXPECTED_ERRORS:
-        logger.error("WCS calculation has failed")
-        raise
-
-    # Update model meta.
-    logger.info("Aperture WCS info: %s", wcsinfo)
-    logger.info("V1 WCS info: %s", vinfo)
-    update_meta(model, t_pars, wcsinfo, vinfo, quality)
-
-    return transforms
-
-
-def calc_wcs_over_time(obsstart, obsend, t_pars: TransformParameters):
-    """
-    Calculate V1 and aperture WCS over a time period.
-
-    Parameters
-    ----------
-    obsstart, obsend : float
-        MJD observation start/end times
-
-    t_pars : `TransformParameters`
-        The transformation parameters. Parameters are updated during processing.
+    position : numpy.array(2)
+        The position in Ideal frame.
 
     Returns
     -------
-    obstimes, wcsinfos, vinfos : [astropy.time.Time[,...]], [WCSRef[,...]], [WCSRef[,...]]
-        A 3-tuple is returned with the WCS pointings for
-        the aperture and the V1 axis.
+    m : np.array(3,3)
+        The transformation matrix
     """
-    # Setup structures
-    obstimes = []
-    wcsinfos = []
-    vinfos = []
+    # Convert to radians
+    ra = wcs.ra * D2R
+    dec = wcs.dec * D2R
+    yangle_ra = yangle * D2R
+    pos_rads = position * A2R
+    v2 = pos_rads[0]
+    v3 = pos_rads[1]
 
-    # Calculate WCS
-    try:
-        pointings = get_pointing(
-            obsstart,
-            obsend,
-            service_kwargs=t_pars.service_kwargs,
-            tolerance=t_pars.tolerance,
-            reduce_func=t_pars.reduce_func,
-        )
-    except ValueError:
-        logger.warning(
-            "Cannot get valid engineering mnemonics from engineering database"
-        )
-        raise
-    if not isinstance(pointings, list):
-        pointings = [pointings]
-    for pointing in pointings:
-        t_pars.pointing = pointing
-        wcsinfo, vinfo, transforms = calc_wcs(t_pars)
-        obstimes.append(pointing.obstime)
-        wcsinfos.append(wcsinfo)
-        vinfos.append(vinfo)
+    # Create the matrices
+    r1 = dcm(ra, dec, yangle_ra)
 
-    return obstimes, wcsinfos, vinfos
-
-
-def calc_wcs(t_pars: TransformParameters):
-    """
-    Calculate WCS.
-
-    Given observatory orientation and target aperture,
-    calculate V1 and Reference Pixel sky coordinates.
-
-    Parameters
-    ----------
-    t_pars : `TransformParameters`
-        The transformation parameters. Parameters are updated during processing.
-
-    Returns
-    -------
-    wcsinfo, vinfo, transforms : WCSRef, WCSRef, Transforms
-        A 3-tuple is returned with the WCS pointing for
-        the aperture and the V1 axis, and the transformation matrices.
-    """
-    # Calculate transforms
-    transforms = calc_transforms(t_pars)
-
-    # Calculate the V1 WCS information
-    vinfo = calc_wcs_from_matrix(transforms.m_eci2v)
-
-    # Calculate the Aperture WCS
-    wcsinfo = wcsinfo_from_siaf(t_pars.aperture, vinfo)
-
-    # That's all folks
-    return wcsinfo, vinfo, transforms
-
-
-def wcsinfo_from_siaf(aperture, vinfo):
-    """Calculate aperture reference point WCS from V-frame WCS and SIAF
-
-    Parameters
-    ----------
-    aperture : str
-        The aperture in use
-
-    vinfo : WCSRef
-        The V-frame WCS
-
-    Returns
-    -------
-    wcsinfo : WCSRef
-        The WCS for the aperture's reference point, as defined by its SIAF.
-    """
-    from pysiaf import Siaf
-    from pysiaf.utils.rotations import sky_posangle
-
-    siaf = Siaf("roman")
-    wfi = siaf[aperture.upper()]
-
-    # For transformations between the telescope frame and all other frames,
-    # an attitude matrix is created using the V-frame WCS information.
-    attitude = attitude_from_v1(vinfo)
-    wfi.set_attitude_matrix(attitude)
-    skycoord = wfi.reference_point(to_frame="sky")
-    pa_v3 = sky_posangle(attitude, *skycoord)
-
-    # Compute S_REGION keyword
-    corners = wfi.corners("sky")
-    footprint = np.array([corners[0], corners[1]]).T
-    s_region_keyword = compute_s_region_keyword(footprint)
-
-    wcsinfo = WCSRef(
-        ra=skycoord[0], dec=skycoord[1], pa=pa_v3, s_region=s_region_keyword
+    r2 = np.array(
+        [
+            [cos(v2) * cos(v3), -sin(v2), -cos(v2) * sin(v3)],
+            [sin(v2) * cos(v3), cos(v2), -sin(v2) * sin(v3)],
+            [sin(v3), 0.0, cos(v3)],
+        ]
     )
-    return wcsinfo
+
+    # Final transformation
+    m = np.dot(r2, r1)
+
+    logger.debug("attitude DCM: %s", m)
+    return m
 
 
 def calc_transforms(t_pars: TransformParameters):
@@ -458,52 +302,6 @@ def calc_transforms(t_pars: TransformParameters):
     return t
 
 
-def calc_attitude_matrix(wcs, yangle, position):
-    """
-    Calculate the DCM attitude from known positions and roll angles.
-
-    Parameters
-    ----------
-    wcs : WCSRef
-        The sky position.
-
-    yangle : float
-        The IdlYangle of the point in question.
-
-    position : numpy.array(2)
-        The position in Ideal frame.
-
-    Returns
-    -------
-    m : np.array(3,3)
-        The transformation matrix
-    """
-    # Convert to radians
-    ra = wcs.ra * D2R
-    dec = wcs.dec * D2R
-    yangle_ra = yangle * D2R
-    pos_rads = position * A2R
-    v2 = pos_rads[0]
-    v3 = pos_rads[1]
-
-    # Create the matrices
-    r1 = dcm(ra, dec, yangle_ra)
-
-    r2 = np.array(
-        [
-            [cos(v2) * cos(v3), -sin(v2), -cos(v2) * sin(v3)],
-            [sin(v2) * cos(v3), cos(v2), -sin(v2) * sin(v3)],
-            [sin(v3), 0.0, cos(v3)],
-        ]
-    )
-
-    # Final transformation
-    m = np.dot(r2, r1)
-
-    logger.debug("attitude DCM: %s", m)
-    return m
-
-
 def calc_wcs_from_matrix(m):
     """
     Calculate the WCS information from a DCM.
@@ -534,6 +332,196 @@ def calc_wcs_from_matrix(m):
 
     logger.debug("wcs: %s", wcs)
     return wcs
+
+
+def calc_wcs_over_time(obsstart, obsend, t_pars: TransformParameters):
+    """
+    Calculate V1 and aperture WCS over a time period.
+
+    Parameters
+    ----------
+    obsstart, obsend : float
+        MJD observation start/end times
+
+    t_pars : `TransformParameters`
+        The transformation parameters. Parameters are updated during processing.
+
+    Returns
+    -------
+    obstimes, wcsinfos, vinfos : [astropy.time.Time[,...]], [WCSRef[,...]], [WCSRef[,...]]
+        A 3-tuple is returned with the WCS pointings for
+        the aperture and the V1 axis.
+    """
+    # Setup structures
+    obstimes = []
+    wcsinfos = []
+    vinfos = []
+
+    # Calculate WCS
+    try:
+        pointings = plib.get_pointing(
+            obsstart,
+            obsend,
+            service_kwargs=t_pars.service_kwargs,
+            tolerance=t_pars.tolerance,
+            reduce_func=t_pars.reduce_func,
+        )
+    except ValueError:
+        logger.warning(
+            "Cannot get valid engineering mnemonics from engineering database"
+        )
+        raise
+    if not isinstance(pointings, list):
+        pointings = [pointings]
+    for pointing in pointings:
+        t_pars.pointing = pointing
+        wcsinfo, vinfo, transforms = calc_wcs(t_pars)
+        obstimes.append(pointing.obstime)
+        wcsinfos.append(wcsinfo)
+        vinfos.append(vinfo)
+
+    return obstimes, wcsinfos, vinfos
+
+
+def calc_wcs(t_pars: TransformParameters):
+    """
+    Calculate WCS.
+
+    Given observatory orientation and target aperture,
+    calculate V1 and Reference Pixel sky coordinates.
+
+    Parameters
+    ----------
+    t_pars : `TransformParameters`
+        The transformation parameters. Parameters are updated during processing.
+
+    Returns
+    -------
+    wcsinfo, vinfo, transforms : WCSRef, WCSRef, Transforms
+        A 3-tuple is returned with the WCS pointing for
+        the aperture and the V1 axis, and the transformation matrices.
+    """
+    # Calculate transforms
+    transforms = calc_transforms(t_pars)
+
+    # Calculate the V1 WCS information
+    vinfo = calc_wcs_from_matrix(transforms.m_eci2v)
+
+    # Calculate the Aperture WCS
+    wcsinfo = wcsinfo_from_siaf(t_pars.aperture, vinfo)
+
+    # That's all folks
+    return wcsinfo, vinfo, transforms
+
+
+def update_wcs_from_telem(model, t_pars: TransformParameters):
+    """
+    Update WCS pointing information.
+
+    Given a `roman.datamodels.DataModel`, determine the simple WCS parameters
+    from the SIAF keywords in the model and the engineering parameters
+    that contain information about the telescope pointing.
+
+    It presumes all the accessed keywords are present.
+
+    Parameters
+    ----------
+    model : `~roman.datamodels.DataModel`
+        The model to update. The update is done in-place.
+
+    t_pars : `TransformParameters`
+        The transformation parameters. Parameters are updated during processing.
+
+    Returns
+    -------
+    transforms : Transforms or None
+        If available, the transformation matrices.
+    """
+    logger.info("Updating wcs from telemetry.")
+
+    # Initialization. If provided, provide a default Pointing.
+    transforms = None  # Assume no transforms are calculated.
+    quality = None  # Unknown pointing quality.
+
+    # Get the pointing information
+    try:
+        t_pars.update_from_engdb()
+    except ValueError as exception:
+        logger.error("Cannot retrieve valid engineering orientation data")
+        if t_pars.default_quaternion is None or not t_pars.allow_default:
+            logger.error("Use of default orientation has been disabled. Aborting.")
+            raise
+        else:
+            logger.warning("Exception is %s", exception)
+            obstime = Time(
+                (t_pars.obsstart.mjd + t_pars.obsend.mjd) / 2.0, format="mjd"
+            )
+            logger.warning(
+                "Using provided default quaternion: %s", t_pars.default_quaternion
+            )
+            logger.warning("    at time %s", obstime.iso)
+            logger.info("Setting pointing quality to PLANNED")
+            t_pars.pointing = plib.Pointing(q=t_pars.default_quaternion, obstime=obstime)
+            quality = "PLANNED"
+    else:
+        logger.info("Successful read of engineering quaternions:")
+        logger.info("\tPointing: %s", t_pars.pointing)
+        quality = "CALCULATED"
+
+    # Attempt to calculate WCS information
+    try:
+        wcsinfo, vinfo, transforms = calc_wcs(t_pars)
+        logger.info("Setting pointing quality to %s", quality)
+    except EXPECTED_ERRORS:
+        logger.error("WCS calculation has failed")
+        raise
+
+    # Update model meta.
+    logger.info("Aperture WCS info: %s", wcsinfo)
+    logger.info("V1 WCS info: %s", vinfo)
+    update_meta(model, t_pars, wcsinfo, vinfo, quality)
+
+    return transforms
+
+
+def wcsinfo_from_siaf(aperture, vinfo):
+    """Calculate aperture reference point WCS from V-frame WCS and SIAF
+
+    Parameters
+    ----------
+    aperture : str
+        The aperture in use
+
+    vinfo : WCSRef
+        The V-frame WCS
+
+    Returns
+    -------
+    wcsinfo : WCSRef
+        The WCS for the aperture's reference point, as defined by its SIAF.
+    """
+    from pysiaf import Siaf
+    from pysiaf.utils.rotations import sky_posangle
+
+    siaf = Siaf("roman")
+    wfi = siaf[aperture.upper()]
+
+    # For transformations between the telescope frame and all other frames,
+    # an attitude matrix is created using the V-frame WCS information.
+    attitude = attitude_from_v1(vinfo)
+    wfi.set_attitude_matrix(attitude)
+    skycoord = wfi.reference_point(to_frame="sky")
+    pa_v3 = sky_posangle(attitude, *skycoord)
+
+    # Compute S_REGION keyword
+    corners = wfi.corners("sky")
+    footprint = np.array([corners[0], corners[1]]).T
+    s_region_keyword = compute_s_region_keyword(footprint)
+
+    wcsinfo = WCSRef(
+        ra=skycoord[0], dec=skycoord[1], pa=pa_v3, s_region=s_region_keyword
+    )
+    return wcsinfo
 
 
 def calc_quat2matrix(q):
@@ -618,81 +606,6 @@ def calc_position_angle(point, ref):
     return point_pa
 
 
-def get_pointing(
-    obsstart,
-    obsend,
-    mnemonics_to_read=COARSE_MNEMONICS,
-    service_kwargs=None,
-    tolerance=60,
-    reduce_func=None,
-):
-    """
-    Get telescope pointing engineering data.
-
-    Parameters
-    ----------
-    obsstart, obsend : float
-        MJD observation start/end times
-
-    mnemonics_to_read : (str,[...])
-        mnemonics to read.
-
-    service_kwargs : dict or None
-        Keyword arguments passed to `engdb_service` defining what
-        engineering database service to use.
-
-    tolerance : int
-        If no telemetry can be found during the observation,
-        the time, in seconds, beyond the observation time to
-        search for telemetry.
-
-    reduce_func : func or None
-        Reduction function to use on values.
-        If None, the average pointing is returned.
-
-    Returns
-    -------
-    pointing : Pointing or [Pointing(, ...)]
-        The engineering pointing parameters.
-        If the `result_type` is `all`, a list
-        of pointings will be returned.
-
-    Raises
-    ------
-    ValueError
-        Cannot retrieve engineering information.
-
-    Notes
-    -----
-    For the moment, the first found values will be used.
-    This will need be re-examined when more information is
-    available.
-    """
-    if reduce_func is None:
-        reduce_func = pointing_from_average
-
-    logger.info("Determining pointing between observations times (mjd):")
-    logger.info("obsstart: %s obsend: %s", obsstart, obsend)
-    logger.info("Telemetry search tolerance: %s", tolerance)
-    logger.info("Reduction function: %s", reduce_func)
-
-    mnemonics = get_mnemonics(
-        obsstart,
-        obsend,
-        mnemonics_to_read=mnemonics_to_read,
-        tolerance=tolerance,
-        service_kwargs=service_kwargs,
-    )
-    reduced = reduce_func(mnemonics)
-
-    logger.log(DEBUG_FULL, "Mnemonics found:")
-    logger.log(DEBUG_FULL, "%s", mnemonics)
-    logger.info("Reduced set of pointings:")
-    logger.info("%s", reduced)
-
-    return reduced
-
-
 def vector_to_angle(v):
     """
     Return tuple of spherical angles from unit direction Vector.
@@ -738,324 +651,6 @@ def angle_to_vector(alpha, delta):
     v2 = sin(delta)
 
     return [v0, v1, v2]
-
-
-def get_mnemonics(
-    obsstart, obsend, tolerance, mnemonics_to_read=COARSE_MNEMONICS, service_kwargs=None
-):
-    """
-    Retrieve pointing mnemonics from the engineering database.
-
-    Parameters
-    ----------
-    obsstart, obsend : float
-        astropy.Time observation start/end times.
-
-    tolerance : int
-        If no telemetry can be found during the observation,
-        the time, in seconds, beyond the observation time to
-        search for telemetry.
-
-    mnemonics_to_read : (str[,...])
-        The mnemonics to fetch.
-
-    service_kwargs : dict or None
-        Keyword arguments passed to `engdb_service` defining what
-        engineering database service to use.
-
-    Returns
-    -------
-    mnemonics : {mnemonic: [value[,...]][,...]}
-        The values for each pointing mnemonic.
-
-    Raises
-    ------
-    ValueError
-        Cannot retrieve engineering information.
-    """
-    if service_kwargs is None:
-        service_kwargs = dict()
-    try:
-        engdb = engdb_service(**service_kwargs)
-    except EXPECTED_ERRORS as exception:
-        raise ValueError(
-            f"Cannot open engineering DB connection\nException: {exception}"
-        ) from exception
-    logger.info("Querying engineering DB: %s", engdb)
-
-    # Construct the mnemonic values structure.
-    mnemonics = {mnemonic: None for mnemonic in mnemonics_to_read}
-
-    # Retrieve the mnemonics from the engineering database.
-    # Check for whether the bracket values are used and
-    # within tolerance.
-    for mnemonic in mnemonics:
-        try:
-            mnemonics[mnemonic] = engdb.get_values(
-                mnemonic,
-                obsstart,
-                obsend,
-                time_format="mjd",
-                include_obstime=True,
-                include_bracket_values=False,
-            )
-        except EXPECTED_ERRORS as exception:
-            logger.warning("Cannot retrieve %s from engineering.", mnemonic)
-            logger.debug("Exception %s", exception)
-            continue
-
-        # If more than two points exist, throw off the bracket values.
-        # Else, ensure the bracket values are within the allowed time.
-        if len(mnemonics[mnemonic]) < 2:
-            logger.warning(
-                "Mnemonic %s has no telemetry within the observation time.", mnemonic
-            )
-            logger.warning(
-                "Attempting to use bracket values within %s seconds", tolerance
-            )
-
-            mnemonics[mnemonic] = engdb.get_values(
-                mnemonic,
-                obsstart,
-                obsend,
-                time_format="mjd",
-                include_obstime=True,
-                include_bracket_values=True,
-            )
-
-            tolerance_mjd = TimeDelta(tolerance, format="sec")
-            allowed_start = obsstart - tolerance_mjd
-            allowed_end = obsend + tolerance_mjd
-            allowed = [
-                value
-                for value in mnemonics[mnemonic]
-                if allowed_start <= value.obstime <= allowed_end
-            ]
-            if not len(allowed):
-                logger.warning(
-                    "No telemetry exists for mnemonic {} within {} and {}".format(
-                        mnemonic,
-                        Time(allowed_start, format="mjd").isot,
-                        Time(allowed_end, format="mjd").isot,
-                    )
-                )
-            mnemonics[mnemonic] = allowed
-
-    return mnemonics
-
-
-def all_pointings(mnemonics):
-    """
-    V1 of making pointings.
-
-    Parameters
-    ----------
-    mnemonics : {mnemonic: [value[,...]][,...]}
-        The values for each pointing mnemonic.
-
-    Returns
-    -------
-    pointings : [Pointing[,...]]
-        List of pointings.
-    """
-    ordered = mnemonics_chronologically(mnemonics)
-    pointings = mnemonics_to_pointings(ordered)
-
-    return pointings
-
-
-def mnemonics_to_pointings(ordered_mnemonics):
-    """From a list of mnemonic values, create a list of Pointings
-
-    Parameters
-    ----------
-    ordered_mnemonics : [(Time, {mnemonic: [value[,...]][,...]})[,...]]
-        List of 2-tuples consisting of (Time, dict) where the dict are
-        the mnemonics values at the associated time.
-
-    Returns
-    -------
-    pointings : [Pointing[,...]]
-        List of pointings
-    """
-    pointings = []
-    for obstime, mnemonics_at_time in ordered_mnemonics:
-        # Observatory orientation, required
-        try:
-            q = np.array(
-                [mnemonics_at_time[m].value for m in COARSE_MNEMONICS_QUATERNION_ECI]
-            )
-        except KeyError as exception:
-            raise ValueError(
-                f"One or more quaternion mnemonics not in the telemetry {COARSE_MNEMONICS_QUATERNION_ECI}"
-            ) from exception
-
-        # B-frame to FGS-frame quaternion. Not required and very oddly has so many backups...
-        fgs_q = None
-        try:
-            fgs_q = np.array(
-                [mnemonics_at_time[m].value for m in COARSE_MNEMONICS_B2FGS_EST]
-            )
-        except KeyError:
-            logger.warning(
-                "One or more of the B-to-FGS quaternion mnemonics are not in the telementry %s",
-                COARSE_MNEMONICS_B2FGS_EST,
-            )
-
-        pointing = Pointing(fgs_q=fgs_q, obstime=obstime, q=q)
-        pointings.append(pointing)
-
-    return pointings
-
-
-def first_pointing(mnemonics):
-    """
-    Return first pointing.
-
-    Parameters
-    ----------
-    mnemonics : {mnemonic: [value[,...]][,...]}
-        The values for each pointing mnemonic.
-
-    Returns
-    -------
-    pointing : Pointing
-        First pointing.
-    """
-    pointings = all_pointings(mnemonics)
-    return pointings[0]
-
-
-def pointing_from_average(mnemonics):
-    """
-    Determine single pointing from average of available pointings.
-
-    Parameters
-    ----------
-    mnemonics : {mnemonic: [value[,...]][,...]}
-        The values for each pointing mnemonic.
-
-    Returns
-    -------
-    pointing : Pointing
-        Pointing from average.
-    """
-    # weed-out empty mnemonics
-    valid_mnemonics = [
-        mnemonic for mnemonic in mnemonics if mnemonics[mnemonic] is not None
-    ]
-
-    # Get average observation time.
-    times = [
-        eng_param.obstime.unix
-        for mnemonic in valid_mnemonics
-        for eng_param in mnemonics[mnemonic]
-        if eng_param.obstime.unix != 0.0
-    ]
-    if len(times) > 0:
-        obstime = Time(np.average(times), format="unix")
-    else:
-        raise ValueError("No valid mnemonics found.")
-
-    # Get averages for all the mnemonics.
-    mnemonic_averages = {}
-    for mnemonic in valid_mnemonics:
-        values = [eng_param.value for eng_param in mnemonics[mnemonic]]
-        if np.allclose(values, 0.0):
-            logger.warning(
-                "Mnemonics %s is only zeros. Treating as undefined.", mnemonic
-            )
-        else:
-            mnemonic_averages[mnemonic] = EngDB_Value(
-                obstime=obstime, value=np.average(values)
-            )
-
-    pointing = mnemonics_to_pointings([(obstime, mnemonic_averages)])[0]
-
-    # That's all folks
-    return pointing
-
-
-def mnemonics_chronologically(mnemonics):
-    """
-    Return time-ordered mnemonic list with progressive values.
-
-    The different set of mnemonics used for observatory orientation
-    appear at different cadences. This routine creates a time-ordered dictionary
-    with all the mnemonics for each time found in the engineering. For mnemonics
-    missing for a particular time, the last previous value is used.
-
-    Parameters
-    ----------
-    mnemonics : {mnemonic: [value[,...]]}
-        Dictionary mapping mnemonics to their respective values.
-
-    Returns
-    -------
-    ordered : [(obstime, {mnemonic: value[,...]}[,...]]
-        Time-ordered list of 2-tuple consisting of `(time, mnemonics)`.
-    """
-    # Collect all information by observation time and sort.
-    by_obstime = defaultdict(dict)
-    for mnemonic, values in mnemonics.items():
-        if values is not None:
-            for value in values:
-                by_obstime[value.obstime][mnemonic] = value
-    by_obstime = sorted(by_obstime.items())
-
-    # Created the ordered matrix
-    ordered = []
-    last_obstime = {}
-    for obstime, mnemonics_at_time in by_obstime:
-        last_obstime.update(mnemonics_at_time)
-
-        # Engineering data may be present, but all zeros.
-        # Filter out this situation.
-        values = [value.value for value in last_obstime.values()]
-        if not any(values):
-            continue
-
-        ordered.append((obstime, copy(last_obstime)))
-
-    return ordered
-
-
-def mnemonics_chronologically_table(mnemonics):
-    """
-    Return time-ordered mnemonic list with progressive values.
-
-    The different set of mnemonics used for observatory orientation
-    appear at different cadences. This routine creates a time-ordered dictionary
-    with all the mnemonics for each time found in the engineering. For mnemonics
-    missing for a particular time, the last previous value is used.
-
-    Parameters
-    ----------
-    mnemonics : {mnemonic: [value[,...]]}
-        Dictionary mapping mnemonics to their respective values.
-
-    Returns
-    -------
-    ordered_by_time : `astropy.table.Table`
-        Time-ordered mnemonic list with progressive values.
-    """
-    ordered = mnemonics_chronologically(mnemonics)
-
-    names = list(mnemonics.keys())
-    names = ["time", *names]
-    time_idx = 0
-
-    values = [[] for _ in names]
-
-    for time, mnemonics_at_time in ordered:
-        values[time_idx].append(time)
-        for mnemonic in mnemonics_at_time:
-            idx = names.index(mnemonic)
-            values[idx].append(ordered[time][mnemonic].value)
-
-    t = Table(values, names=names)
-
-    return t
 
 
 def t_pars_from_model(model, t_pars):
