+++ romancal/orientation/tests/test_set_telescope_pointing.py
@@ -15,6 +15,7 @@ from astropy.time import Time
 from romancal.lib.engdb import engdb_mast, engdb_tools
 from romancal.orientation import set_telescope_pointing as stp
 from romancal.orientation import _lib as olib
+from romancal.orientation import _pointing as plib
 
 # pysiaf is not a required dependency. If not present, ignore all this.
 pysiaf = pytest.importorskip("pysiaf")
@@ -41,7 +42,7 @@ TARG_DEC = 66.0
 TRANSFORM_KWARGS = {
     "aperture": "WFI01_FULL",
     "gscommanded": (916.4728835141, -186.8939737044),
-    "pointing": olib.Pointing(
+    "pointing": plib.Pointing(
         fgs_q=np.array(
             [
                 -0.18596734175399293,
@@ -124,7 +125,7 @@ def test_change_base_url():
     """
     service_kwargs = {"service": "mast", "eng_base_url": engdb_mast.MAST_BASE_URL}
     with pytest.raises(ValueError):
-        olib.get_pointing(
+        plib.get_pointing(
             Time("2015-06-15"), Time("2015-06-17"), service_kwargs=service_kwargs
         )
 
@@ -136,7 +137,7 @@ def test_change_base_url_fail():
         "eng_base_url": "https://nonexistent.fake.example",
     }
     with pytest.raises(ValueError):
-        olib.get_pointing(
+        plib.get_pointing(
             Time(STARTTIME, format="isot"),
             Time(ENDTIME, format="isot"),
             service_kwargs=service_kwargs,
@@ -147,11 +148,11 @@ def test_change_base_url_fail():
 def test_get_mnemonics():
     """Test getting mnemonics"""
     try:
-        mnemonics = olib.get_mnemonics(STARTTIME, ENDTIME, 60)
+        mnemonics = plib.get_mnemonics(STARTTIME, ENDTIME, 60)
     except ValueError as exception:
         pytest.xfail(reason=str(exception))
 
-    assert len(mnemonics) == len(olib.COARSE_MNEMONICS)
+    assert len(mnemonics) == len(plib.COARSE_MNEMONICS)
 
 
 @pytest.mark.skipif(NO_ENGDB, reason="No engineering database available")
@@ -162,7 +163,7 @@ def test_get_pointing():
     This will most likely change again.
     """
     try:
-        pointing = olib.get_pointing(STARTTIME, ENDTIME)
+        pointing = plib.get_pointing(STARTTIME, ENDTIME)
     except ValueError as exception:
         pytest.xfail(reason=str(exception))
 
@@ -174,7 +175,7 @@ def test_get_pointing():
 
 def test_get_pointing_fail():
     with pytest.raises(ValueError):
-        obstime, q = olib.get_pointing(BADSTARTTIME, BADENDTIME)
+        obstime, q = plib.get_pointing(BADSTARTTIME, BADENDTIME)
 
 
 @pytest.mark.skipif(NO_ENGDB, reason="No engineering database available")
@@ -185,7 +186,7 @@ def test_get_pointing_list():
     This will most likely change again.
     """
     try:
-        results = olib.get_pointing(STARTTIME, ENDTIME, reduce_func=olib.all_pointings)
+        results = plib.get_pointing(STARTTIME, ENDTIME, reduce_func=plib.all_pointings)
     except ValueError as exception:
         pytest.xfail(reason=str(exception))
     assert isinstance(results, list)
@@ -212,23 +213,23 @@ def test_hv_to_fgs():
 def test_mnemonics_chronologically():
     """Test ordering mnemonics chronologically"""
     try:
-        mnemonics = olib.get_mnemonics(STARTTIME, ENDTIME, 60)
+        mnemonics = plib.get_mnemonics(STARTTIME, ENDTIME, 60)
     except ValueError as exception:
         pytest.xfail(reason=str(exception))
-    ordered = olib.mnemonics_chronologically(mnemonics)
+    ordered = plib.mnemonics_chronologically(mnemonics)
 
     assert len(ordered) > 1
 
     first = ordered[0]
     assert isinstance(first[0], Time)
     assert isinstance(first[1], dict)
-    assert len(first[1]) >= len(olib.COARSE_MNEMONICS_QUATERNION_ECI)
+    assert len(first[1]) >= len(plib.COARSE_MNEMONICS_QUATERNION_ECI)
 
 
 @pytest.mark.skipif(NO_ENGDB, reason="No engineering database available")
 def test_logging(caplog):
     try:
-        olib.get_pointing(STARTTIME, ENDTIME)
+        plib.get_pointing(STARTTIME, ENDTIME)
     except ValueError as exception:
         pytest.xfail(reason=str(exception))
     assert "Determining pointing between observations times" in caplog.text
@@ -252,7 +253,7 @@ def test_mnemonic_list():
         )
     )
 
-    assert expected == set(olib.COARSE_MNEMONICS)
+    assert expected == set(plib.COARSE_MNEMONICS)
 
 
 @pytest.mark.parametrize("wcs_type", ["wcsinfo", "vinfo"])
